Changes made for SaveFile() function
Fully commented the SaveFile() function.
Added a check at the beginning of the function, where if there exists a file at the path provided by <path>, then the user will be shown a prompt where they can choose whether or not they wish to overwrite the file.... Why? well this is to try and attempt to stop the program from accidentally overwriting the old file.